+++ spec/regression/test_0002.rb
@@ -1,0 +1,12 @@
+require 'test_helpers'
+describe Alf do
+
+  it 'should support projecting over no attribute' do
+    db.connect do |conn|
+      rel = conn.query{ project(suppliers, []) }
+      exp = Relation::DEE
+      rel.should eq(exp)
+    end
+  end
+
+end