+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # 0.15.1 - FIX ME
 
 * Fixed Tuple() that now accepts no arg and returns an empty tuple.
+* Fixed SQL compilation of join with DEE/DUM.
+* Removed `is_table_dee` attribute when projecting all attributes away from
+  a SQL data source.
 
 # 0.15.0 - 2013-11-01